chore(tests): rename fallback to native
